feat(Produtos): Boas práticas de implementação

+++ Controllers/ProdutosController.cs
@@ -18,22 +18,44 @@ public class ProdutosController : ControllerBase
     }
 
     [HttpGet]
+    [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesDefaultResponseType]
     public async Task<ActionResult<IEnumerable<Produto>>> GetProdutos()
     {
+        if (_context.Produtos is null)
+            return NotFound();
+
         return Ok(await _context.Produtos.ToListAsync());
     }
 
     [HttpGet("{id:Guid}")]
+    [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesDefaultResponseType]
     public async Task<ActionResult<Produto>> GetProduto(Guid id)
     {
+        if (_context.Produtos is null)
+            return NotFound();
+
         var produto = await _context.Produtos.FindAsync(id);
+
+        if (produto is null)
+            return NotFound();
 
         return Ok(produto);
     }
 
     [HttpPost]
+    [ProducesResponseType(StatusCodes.Status201Created)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
+    [ProducesDefaultResponseType]
     public async Task<ActionResult<Produto>> PostProduto(Produto produto)
     {
+        if (_context.Produtos is null)
+            return Problem("Erro ao criar um produto, contate o suporte!");
+
+
         if (!ModelState.IsValid)
             return ValidationProblem(new ValidationProblemDetails(ModelState)
             {
@@ -47,6 +69,10 @@ public class ProdutosController : ControllerBase
     }
 
     [HttpPut("{id:Guid}")]
+    [ProducesResponseType(StatusCodes.Status204NoContent)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesDefaultResponseType]
     public async Task<ActionResult<Produto>> PutProduto(Guid id, Produto produto)
     {
         if (id != produto.Id)
@@ -55,20 +81,45 @@ public class ProdutosController : ControllerBase
         if (!ModelState.IsValid)
             return ValidationProblem(ModelState);
 
-        _context.Produtos.Update(produto);
-        await _context.SaveChangesAsync();
+        _context.Entry(produto).State = EntityState.Modified;
+
+        try
+        {
+            await _context.SaveChangesAsync();
+        }
+        catch (DbUpdateConcurrencyException)
+        {
+            if (!ProdutoExists(id))
+                return NotFound();
+            else
+                throw;
+        }
 
         return NoContent();
     }
 
     [HttpDelete("{id:Guid}")]
+    [ProducesResponseType(StatusCodes.Status204NoContent)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesDefaultResponseType]
     public async Task<ActionResult<Produto>> DeleteProduto(Guid id)
     {
+        if (_context.Produtos is null)
+            return NotFound();
+
         var produto = await _context.Produtos.FindAsync(id);
+
+        if (produto is null)
+            return NotFound();
 
         _context.Produtos.Remove(produto);
         await _context.SaveChangesAsync();
 
         return NoContent();
+    }
+
+    private bool ProdutoExists(Guid id)
+    {
+        return (_context.Produtos?.Any(e => e.Id == id)).GetValueOrDefault();
     }
 }
